fix value class and size type to video

+++ stream/src/main/kotlin/com/cherhy/plugins/domain/Video.kt
@@ -4,7 +4,6 @@ import com.cherhy.common.util.model.UserId
 import com.cherhy.plugins.util.model.BaseEntity
 import com.cherhy.plugins.util.model.BaseEntityFactory
 import com.cherhy.plugins.util.model.BaseTable
-import org.ktorm.schema.int
 import org.ktorm.schema.long
 import org.ktorm.schema.varchar
 
@@ -12,7 +11,7 @@ object Videos : BaseTable<Video>("video") {
     val id = long("id").primaryKey().bindTo { it.id.value }
     val name = varchar("name").bindTo { it.name.value }
     val uniqueName = varchar("unique_name").bindTo { it.uniqueName.value }
-    val size = int("size").bindTo { it.size.value }
+    val size = long("size").bindTo { it.size.value }
     val extension = varchar("extension").bindTo { it.extension.value }
     val owner = long("owner").bindTo { it.owner.value }
     val post = long("post").bindTo { it.post.value }
@@ -43,12 +42,12 @@ value class VideoId(
 
 @JvmInline
 value class VideoSize(
-    val value: Int,
+    val value: Long,
 ) {
     companion object {
         @JvmStatic
         fun of(
-            value: Int,
+            value: Long,
         ) = VideoSize(value)
     }
 }
